WellnessBot/requirements.txt
aiogram>=3.3.0
aiohttp-socks>=0.10.1
httpx>=0.27.0
openai>=1.0.0
python-dotenv>=1.0.1
jinja2>=3.1.3
playwright>=1.41.2
pypdf>=4.0.0
